+++ source/com/bombinggames/caveland/GameObjects/Vanya.java
@@ -268,7 +268,7 @@ public class Vanya extends MovableEntity implements Interactable, Telegraph {
 					text
 				);
 				if (chatCounter > 0) {
-					currentChat.setConfirmAction((int result, AbstractEntity actor1) -> {
+					currentChat.setConfirmAction((byte result, AbstractEntity actor1) -> {
 						nextChat(view, actor, true);
 					});
 					currentChat.setCancelAction((int result, AbstractEntity actor1) -> {
